ILLDAN-190 [refactor]: kakao 요청에 대한 response의 패키지 구조를 변경한다
