+++ src/main/java/server/poptato/infra/oauth/kakao/KakaoApiClient.java
@@ -6,6 +6,7 @@ import org.springframework.context.annotation.ComponentScan;
 import org.springframework.http.HttpHeaders;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.RequestHeader;
+import server.poptato.infra.oauth.kakao.response.KakaoUserResponse;
 
 
 @ComponentScan
