+++ src/main/java/server/poptato/infra/oauth/kakao/response/KakaoTokenResponse.java
@@ -1,4 +1,4 @@
-package server.poptato.infra.oauth.kakao;
+package server.poptato.infra.oauth.kakao.response;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
